requirements.txt
absl-py==0.15.0
aiohttp==3.8.0
aiosignal==1.2.0
antlr4-python3-runtime==4.8
async-timeout==4.0.0
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
fairscale==0.4.1
frozenlist==1.2.0
fsspec==2021.10.1
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.41.1
hydra-core==1.1.1
idna==3.3
importlib-resources==5.4.0
Markdown==3.3.4
mock==4.0.3
multidict==5.2.0
numpy==1.21.3
oauthlib==3.1.1
omegaconf==2.1.1
packaging==21.2
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
Pygments==2.10.0
pyparsing==2.4.7
pytorch-lightning @ git+https://github.com/PyTorchLightning/pytorch-lightning.git@fff491dd4d42adbbb4bb87aa28e96810b0c2be4d
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
TestSlide==2.7.0
torch==1.10.0
torchmetrics==0.6.0
tqdm==4.62.3
typeguard==2.13.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
zipp==3.6.0
